gh-693: (refactor) extract repition from {local,network} getmeta
