+++ graphqlapi/common/getmeta/property_numerical.go
@@ -15,51 +15,55 @@ package getmeta
 import (
 	"fmt"
 
+	"github.com/creativesoftwarefdn/weaviate/graphqlapi/common"
 	"github.com/creativesoftwarefdn/weaviate/graphqlapi/descriptions"
 	"github.com/creativesoftwarefdn/weaviate/models"
 	"github.com/graphql-go/graphql"
 )
 
-func numberPropertyFields(class *models.SemanticSchemaClass,
-	property *models.SemanticSchemaClassProperty) *graphql.Object {
+func numericalPropertyField(class *models.SemanticSchemaClass,
+	property *models.SemanticSchemaClassProperty, prefix string) *graphql.Object {
 	getMetaNumberFields := graphql.Fields{
 		"sum": &graphql.Field{
-			Name:        fmt.Sprintf("Meta%s%sSum", class.Class, property.Name),
+			Name:        fmt.Sprintf("%s%s%sSum", prefix, class.Class, property.Name),
 			Description: descriptions.GetMetaPropertySumDesc,
 			Type:        graphql.Float,
+			Resolve:     common.JSONNumberResolver,
 		},
 		"type": &graphql.Field{
-			Name:        fmt.Sprintf("Meta%s%sType", class.Class, property.Name),
+			Name:        fmt.Sprintf("%s%s%sType", prefix, class.Class, property.Name),
 			Description: descriptions.GetMetaPropertyTypeDesc,
 			Type:        graphql.String,
 		},
 		"lowest": &graphql.Field{
-			Name:        fmt.Sprintf("Meta%s%sLowest", class.Class, property.Name),
+			Name:        fmt.Sprintf("%s%s%sLowest", prefix, class.Class, property.Name),
 			Description: descriptions.GetMetaPropertyLowestDesc,
 			Type:        graphql.Float,
+			Resolve:     common.JSONNumberResolver,
 		},
 		"highest": &graphql.Field{
-			Name:        fmt.Sprintf("Meta%s%sHighest", class.Class, property.Name),
+			Name:        fmt.Sprintf("%s%s%sHighest", prefix, class.Class, property.Name),
 			Description: descriptions.GetMetaPropertyHighestDesc,
 			Type:        graphql.Float,
+			Resolve:     common.JSONNumberResolver,
 		},
 		"average": &graphql.Field{
-			Name:        fmt.Sprintf("Meta%s%sAverage", class.Class, property.Name),
+			Name:        fmt.Sprintf("%s%s%sAverage", prefix, class.Class, property.Name),
 			Description: descriptions.GetMetaPropertyAverageDesc,
 			Type:        graphql.Float,
+			Resolve:     common.JSONNumberResolver,
 		},
 		"count": &graphql.Field{
-			Name:        fmt.Sprintf("Meta%s%sCount", class.Class, property.Name),
+			Name:        fmt.Sprintf("%s%s%sCount", prefix, class.Class, property.Name),
 			Description: descriptions.GetMetaPropertyCountDesc,
 			Type:        graphql.Int,
+			Resolve:     common.JSONNumberResolver,
 		},
 	}
 
-	getMetaNumberProperty := graphql.ObjectConfig{
-		Name:        fmt.Sprintf("Meta%s%sObj", class.Class, property.Name),
+	return graphql.NewObject(graphql.ObjectConfig{
+		Name:        fmt.Sprintf("%s%s%sObj", prefix, class.Class, property.Name),
 		Fields:      getMetaNumberFields,
 		Description: descriptions.GetMetaPropertyObjectDesc,
-	}
-
-	return graphql.NewObject(getMetaNumberProperty)
+	})
 }
