+++ graphqlapi/common/resolver.go
@@ -10,7 +10,7 @@
  * See www.creativesoftwarefdn.org for details
  * Contact: @CreativeSofwFdn / bob@kub.design
  */
-package common_resolver
+package common
 
 import (
 	"errors"
