+++ graphqlapi/local/get/get_test.go
@@ -18,7 +18,7 @@ import (
 	"testing"
 
 	"github.com/creativesoftwarefdn/weaviate/database/schema/kind"
-	"github.com/creativesoftwarefdn/weaviate/graphqlapi/common_resolver"
+	"github.com/creativesoftwarefdn/weaviate/graphqlapi/common"
 	test_helper "github.com/creativesoftwarefdn/weaviate/graphqlapi/test/helper"
 )
 
@@ -66,7 +66,7 @@ func TestExtractPagination(t *testing.T) {
 		Kind:       kind.ACTION_KIND,
 		ClassName:  "SomeAction",
 		Properties: []SelectProperty{{Name: "intField", IsPrimitive: true}},
-		Pagination: &common_resolver.Pagination{
+		Pagination: &common.Pagination{
 			First: 10,
 			After: 20,
 		},
